+++ plugin/android-junit5/src/main/kotlin/de/mannodermaus/gradle/plugins/junit5/tasks/AndroidJUnit5WriteFilters.kt
@@ -1,15 +1,15 @@
-@file:Suppress("DEPRECATION")
-
 package de.mannodermaus.gradle.plugins.junit5.tasks
 
+import com.android.build.api.variant.SourceDirectories
 import com.android.build.api.variant.Variant
-import com.android.build.gradle.api.TestVariant
 import de.mannodermaus.gradle.plugins.junit5.internal.config.INSTRUMENTATION_FILTER_RES_FILE_NAME
 import de.mannodermaus.gradle.plugins.junit5.internal.config.JUnit5TaskConfig
 import de.mannodermaus.gradle.plugins.junit5.internal.extensions.getTaskName
 import de.mannodermaus.gradle.plugins.junit5.internal.extensions.junitPlatform
 import org.gradle.api.DefaultTask
 import org.gradle.api.Project
+import org.gradle.api.file.DirectoryProperty
+import org.gradle.api.provider.ListProperty
 import org.gradle.api.tasks.CacheableTask
 import org.gradle.api.tasks.Input
 import org.gradle.api.tasks.OutputDirectory
@@ -25,20 +25,20 @@ import java.io.File
  * This only allows tests to be filtered with @Tag annotations even in the instrumentation test realm.
  * Other plugin DSL settings, like includeEngines/excludeEngines or includePattern/excludePattern
  * are not copied out to file. This has to do with limitations of the backport implementation
- * of the JUnit Platform Runner, as well as some incompatibilities between Gradle and Java with regards to
- * how class name patterns are formatted.
+ * of the JUnit Platform Runner, as well as some incompatibilities between Gradle and Java
+ * regarding how class name patterns are formatted.
  */
 @CacheableTask
 public abstract class AndroidJUnit5WriteFilters : DefaultTask() {
 
     internal companion object {
+        @Suppress("UnstableApiUsage")
         fun register(
             project: Project,
             variant: Variant,
-            instrumentationTestVariant: TestVariant
+            sourceDirs: SourceDirectories.Layered,
         ): Boolean {
-            val outputFolder = File("${project.buildDir}/generated/res/android-junit5/${instrumentationTestVariant.name}")
-            val configAction = ConfigAction(project, variant, outputFolder)
+            val configAction = ConfigAction(project, variant)
 
             val provider = project.tasks.register(
                 configAction.name,
@@ -48,54 +48,53 @@ public abstract class AndroidJUnit5WriteFilters : DefaultTask() {
 
             // Connect the output folder of the task to the instrumentation tests
             // so that they are bundled into the built test application
-            instrumentationTestVariant.registerGeneratedResFolders(
-                project.files(outputFolder).builtBy(provider)
+            sourceDirs.addGeneratedSourceDirectory(
+                taskProvider = provider,
+                wiredWith = AndroidJUnit5WriteFilters::outputFolder,
             )
-            instrumentationTestVariant.mergeResourcesProvider.configure { it.dependsOn(provider) }
 
             return true
         }
     }
 
-    @Input
-    public var includeTags: List<String> = emptyList()
+    @get:Input
+    public abstract val includeTags: ListProperty<String>
 
-    @Input
-    public var excludeTags: List<String> = emptyList()
+    @get:Input
+    public abstract val excludeTags: ListProperty<String>
 
-    @OutputDirectory
-    public var outputFolder: File? = null
+    @get:OutputDirectory
+    public abstract val outputFolder: DirectoryProperty
 
     @TaskAction
     public fun execute() {
-        this.outputFolder?.let { folder ->
-            // Clear out current contents of the generated folder
-            folder.deleteRecursively()
+        // Clear out current contents of the generated folder
+        val folder = outputFolder.get().asFile
+        folder.deleteRecursively()
 
-            if (this.hasFilters()) {
-                folder.mkdirs()
+        val includeTags = includeTags.get()
+        val excludeTags = excludeTags.get()
 
-                // Re-write the new file structure into it;
-                // the generated file will have a fixed name & is located
-                // as a "raw" resource inside the output folder
-                val rawFolder = File(folder, "raw").apply { mkdirs() }
-                File(rawFolder, INSTRUMENTATION_FILTER_RES_FILE_NAME)
-                    .bufferedWriter()
-                    .use { writer ->
-                        // This format is a nod towards the real JUnit 5 ConsoleLauncher's arguments
-                        includeTags.forEach { tag -> writer.write("-t $tag") }
-                        excludeTags.forEach { tag -> writer.write("-T $tag") }
-                    }
-            }
+        if (includeTags.isNotEmpty() || excludeTags.isNotEmpty()) {
+            folder.mkdirs()
+
+            // Re-write the new file structure into it;
+            // the generated file will have a fixed name & is located
+            // as a "raw" resource inside the output folder
+            val rawFolder = File(folder, "raw").apply { mkdirs() }
+            File(rawFolder, INSTRUMENTATION_FILTER_RES_FILE_NAME)
+                .bufferedWriter()
+                .use { writer ->
+                    // This format is a nod towards the real JUnit 5 ConsoleLauncher's arguments
+                    includeTags.forEach { tag -> writer.appendLine("-t $tag") }
+                    excludeTags.forEach { tag -> writer.appendLine("-T $tag") }
+                }
         }
     }
-
-    private fun hasFilters() = includeTags.isNotEmpty() || excludeTags.isNotEmpty()
 
     private class ConfigAction(
         private val project: Project,
         private val variant: Variant,
-        private val outputFolder: File
     ) {
 
         val name: String = variant.getTaskName(prefix = "writeFilters", suffix = "androidTest")
@@ -103,12 +102,12 @@ public abstract class AndroidJUnit5WriteFilters : DefaultTask() {
         val type = AndroidJUnit5WriteFilters::class.java
 
         fun execute(task: AndroidJUnit5WriteFilters) {
-            task.outputFolder = outputFolder
-
-            // Access filters for this particular variant & provide them to the task, too
+            // Access filters for this particular variant & provide them to the task
             val configuration = JUnit5TaskConfig(variant, project.junitPlatform)
-            task.includeTags = configuration.combinedIncludeTags.toList()
-            task.excludeTags = configuration.combinedExcludeTags.toList()
+            task.includeTags.set(configuration.combinedIncludeTags.toList())
+            task.excludeTags.set(configuration.combinedExcludeTags.toList())
+
+            // Output folder is applied by Android Gradle Plugin, so there is no reason to provide a value ourselves
         }
     }
 }